cli/requirements.txt
typer
rich
textual
typer
pyyaml
requests